+++ src/main/kotlin/org/jetbrains/plugins/template/services/MyProjectService.kt
@@ -1,12 +1,15 @@
 package org.jetbrains.plugins.template.services
 
+import com.intellij.openapi.components.Service
+import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
 import org.jetbrains.plugins.template.MyBundle
 
+@Service(Service.Level.PROJECT)
 class MyProjectService(project: Project) {
 
     init {
-        println(MyBundle.message("projectService", project.name))
+        thisLogger().info(MyBundle.message("projectService", project.name))
 
         System.getenv("CI")
             ?: TODO("Don't forget to remove all non-needed sample code files with their corresponding registration entries in `plugin.xml`.")